Basic Recall System in!!!!

+++ Assets/Scripts/Enemy/MeleeEnemyAttack.cs
@@ -14,7 +14,7 @@ public class MeleeEnemyAttack : MonoBehaviour
     private VoidBoolCallback isBlocked;
     private bool initialized = false;
     
-    private float _attackPower;
+    private int _attackPower;
 
 
     private OperatorAttack attack;
@@ -32,7 +32,7 @@ public class MeleeEnemyAttack : MonoBehaviour
         }
     }
 
-    public void Initialize(float attackPower)
+    public void Initialize(int attackPower)
     {
         _attackPower = attackPower;
         initialized = true;
